backend/requirements.txt

appdirs==1.4.3

asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4

certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3

cryptography==2.8



Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-passwords==0.3.12
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0

entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.2
keyring==18.0.1

launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2

macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycairo==1.16.2


PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1

python-dateutil==2.7.3

pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
stevedore==1.32.0

testresources==2.0.1


virtualenv==20.0.17
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.3

zope.event==4.5.0
zope.interface==5.4.0